refactor: replace existing pages
